Significantly expanded the auth-test application. There is now a working authentication and authorization framework, although only for form authentication at this time.
* Added resource to view all permissions.

* Added resources to add and delete users.

* Added resources to add and delete roles

As a default, each HTTP method on a specific resource has its own
permission, although this can of course easily be overridden. A role
is a set of permissions, and may in the future also be able to include
other roles. Each user can have multiple roles, and additionally,
multiple permissions as well. There is one special permission, "all",
which as expected implies having all permissions.

The users and roles are stored in tables in the internal database,
which is essentially just a glorified set of dictionaries, but with
some added convenience methods on top. As a default, an application
runs with an in-memory database, which prevents created users and
roles from being persisted when the application exits. To become
really useful, the application needs to be started with a database
file name option. See the online help for more information.